refactor: move common methods
